refactoring to example view
