--- apps/sampleapp/README.md
@@ -1,3 +1,0 @@
-# Example Application based on Deployment
-
-Simple way to install and uninstall a Kubernetes deployment.
