+++ example/README.md
@@ -1,0 +1,17 @@
+# Examples
+
+Use these examples for your application.
+
+### Simple
+
+A point and click example that uses the default command line tools.
+
+Copy it.
+
+Type `make`
+
+Type `sudo make install`.
+
+### Controller
+
+Todo! Pull requests accepted!
