+++ literal.go
