--- pkg/README.md
@@ -1,6 +1,0 @@
-# Package
-
-naml core libraries. 
-
-Most of this is just "fluff" that wraps up `client-go` to make it easier to manage sets of "applications" and test them.
-
